+++ Sophia_Chernousova/resources/rosalind_dna.txt
@@ -1,0 +1,1 @@
+CGCTTTTCCATGACTGGGTCGAAAATGAGATTGATTCACCGCTTCATTATGTCGAGCTCCCATGAGTAGGAGTAACCCCGGTGAACTGAGTACAGCATTTGAAAGCTAGTCCGGTCGAGCGTGGTTGAACTCAAGAGATAGCGGGACAGGTAATTCCCGCTTGTTATCACCCAGGCCTGATCCTGTTGACGGGCTGCGGCCGGTTTACGTCTCGACCGTCTTCAGGGTCAATGTGCTATAAGTATAGGTGCATCAAAAGCACTGCATTTAAAGCGCTCGAAACACGAACAAAGGTGCACTGATGTATTTATCCACCAGGGCGAGCAACTCTTTAGAAATCGGCCCCATCAAATGAGTGCTAGCTGGACTTCACACTCGTTGGATGGACCACCTCCGTGGAGTTCGGACCGGGGAGCCGGTATGCTATCGTTATTTGTGGTACGCTCTACAGTGAGGATCCAATACAGGTTCCGGACATTCCCAAATCACGGAATCAGTGCAAGACATAACATCTTGCTGCCTTGAGGTAGGGACTGAACGCAAACTTGACGGGCCTCGGGGAACGGTGCTACAGTCCCATGATCCTGCAACGCTTACCGGGAGATTGAACTAGTGAGTAGGAACTCTAAGCGTCGTATAAGGTAGTCTTTAGGGACAGATTGGCATCGTATCAACGGTCGTGACCAATCGACTTACAAGAACCGTTCACTTGTCTCTTGTCAGGCGTTTGGTAAGTCTTCTCAGCCTGTGCGCGCACGCCGACTTCTTTCATAAACTTAAACGGGAGTAAATCGGAGATCGCTCTTGACTAAAGAACCTGGGGGT
