+++ Sophia_Chernousova/resources/rosalind_fib.txt
@@ -1,0 +1,1 @@
+35 4
