+++ Sophia_Chernousova/resources/rosalind_hamm.txt
@@ -1,0 +1,2 @@
+GGTAGGGGTGAGACTTTCGCAGGGTGTGAGGAGGCCAAGAGACACAAAGCCTCGGAGGATGCTGTCCGACTGTCGTAAGAGTAGAAGCGGCCAACAGTCACACAAACCACTGCGGCGACAAGAGGTTTCAATCCTCTTAGCGAGTCTCATAACCACATCAGGCCCTAAAGGTGCGCTTAACGTACAAACGCTAATACCGAATAGTAACGGAACAATGCAGATTGCAGTGAGCCGTTTTAGTTCCCGTGTCCCGGCGCTTGGTATTCTCTAACCATTGCTGAATATGACACCTCACTGAGTGCAGCACGAAATTCACATCGCCGCGGACGGGAATCGATCAACCGCCGCGAATTGAGTTTGTTCAGGACTTGCCGACAACAATAGAGCTATAGACAGGGGTATCAGACCACTCGGGGAATCCATCTCCCCAAACTCTCCTCCGAAGTTTGACGCTAGGTTCGGGTACGGGTTAGTAAACTCGTCTGCCAGTAAAGTGGTCCCTGCGCTCCCGCGTAAATGAGCCAGGTGTCTAATTCGTTTGGACCGCCAGCTCGTTGTCTTCTGGTAATGGCAAGCTTTTGTGCGCTCACGTGCTCATGGCGTCACGAATCTTCCCCCATTGGCATAAAGTGATGGAGAGAGAGGAAGTCTACAAACGGTCCGATGTTTTTTCGACTACAACCCATCGGCCGCCTGCCTAGAGACCACCCCACTCTGGGGCCCAGGGTCTTTACTGAATGCGGGACGCACGCGCAACACATCTGAGCATAGAGTACCTCGCTTTTAGCCATCAGGAGTTCCTGTTGGGGAACCCGTTAATTCCCAGACAAATTACATTCTTTTACTGTCAACCATTTTGCACTGGGACGCACCGGGTTTAAAGAGAACTGCCGCACTGCGGACTCGACGCCGCAATGCCCTACTTCCATCT
+GGTATATATGTCGCCTAAGACGGGATAGGGGGGAGCGAAGGTACCTAAGGCTGCCATGATACTGTCAGATGGGTCTAAACGCCGCAGCGGGCCTATTGCGTCCCGAGGTAACCGTCTTCAAGACGTTCAAATAGATCTAGCCTGACTCATCCGTCCACCAGGGCGCATACGTGCCCTAGGATTAGGGGGGCGGAACACGTAAACTCTGGGACAGTACATTCTTACCAATAGGCAATCAAGACCTAGCGGCCCGGTGGCGGATATTAAGTTGGCGTCGCTCTCTGCTGCGCTTCATTAACTGCGGCTCGGAATTCCACTCTTCGCGGGTGGCCGCCAGGAACTGGTGCAGAATTGAGTCATCTACCGAAGTACCGGATCAAAGAGAGCTGTCCACAGGGCCCTCAAAAAGCTACGGCGACCCTCATGTCATAACTTAGGGCCGTAGGTTGACGCTAGCACCGGGTGATGACCGGTCAGCTCAGCTGCGTCGTAAAAAAGCCAAGTGGTCACCAGTATCTTATACTGGAGTCCACTTGGTGTGGAACGGGAGTACAGTAGCTTCTAGAAATCGGGAGCGGGTCTGCGGTCAAGATGTGCTGGCGTTACTCCGCTTCCGCCGTTCACATATAAGTAGCGATCGAGAACCATTTTTACGTGAATCGGATGTTAAAGCGCCTGCAAGTCATTGCCCCCAACGATAGCAAGTGCGCAGCTCTAGGGCCTAGTGACTTTACCAAACGAGAGACTTAAGCGGAACAGACTTGAGCACGCGGTGCCTGGGATTCCGCTTTGACCGGTCCGTTTTGTGGGACCCGTCGAATCACACTAAAATTATAACGTTTTTCGGACCCCGATTGAAGACTGACACGCTACTGGTGCAAAGAGAATTGAAATCATGTGGTGTAGGCCCATCAATGCCTCTTTGCACTAT
