+++ Sophia_Chernousova/resources/rosalind_iprb.txt
@@ -1,0 +1,1 @@
+16 16 17
