+++ Sophia_Chernousova/resources/rosalind_perm.txt
@@ -1,0 +1,1 @@
+6
