+++ Sophia_Chernousova/resources/rosalind_prtm.txt
@@ -1,0 +1,1 @@
+NPWSCRFSQHWQEIAFLTLRLQELYGEDECMWRAIQEWDAGRCYYANPPHHMDARDCWLHSTQASVMKRMHAHKIEMGCCALHRCQPPMTYFVRVCQGTMWAWLELEGHLDCFPLRPPIGANYPMKYWLFFVQDQMIYCNEDTNGVIWPHWPYISKTVMKDSICHELNWGWSYWWDWPPTHFTVWAYTQVMLNWTKEELRDADEMDTQPLVWSHCRVCETEICKAHLAMLQYTDNLSHQFGGMMHYAAEKHAQTNESYDIDMHDIMLCAIQVYGLWDHLMFTYSIRPTPPKQATYFMTPCKPIEDMIISCCQSFVVSLRINMAYGKLNLGHCAPTTNSGFPNYGMNTRNAHWLEWLFDTEACQEHCRWAESICPKYCVASFRQPFNLGRAVYGYEGCDAVPAEGCHAFYAEMISWTGDCDQVIRRHWQCLEMNRWHKHAGRKDLAYCTYTWPKAPFGERNQCNERLDKGDAYVEFSRCANLARASLVSVAGNIKWKILQVCFIKRCMSSIYEQTVYRASQPDAMTSAPCWFWILTDQWGLTNETMYSWPYLPSWQKTEDQRASQEWGAGNAPTHKSLIQGMEGHSFKVSFLEENYRGHHSCAEVTPEDVNHTCGITKKEMAPPFSWWECKIPSFKLCELPWAEEHMELLFMMCLIWAWCCELKQDHWNHWESPMCRSFLPDAYCVWDVWHDPECQKIKQADIQGNHPATDIDAAEVYVRTTERQAWHSMEPTNKCIKYCKKIYVQWQASIEKSPHQPQCAHTYVTARMGIDAIKHQFVIQSEITMLHLKYTQECFKKTMLHVMIENISQGLFVAGWMFQWIHLNNYHDPLPSMRAKVHLAVWDASPFRCYAQNQKWYFPNNNVAFIFMWNEHWCTAVEQRDDNEPFKIPLCEMAQDRWCNRYFDYDGEFLVLFPNSTRGCQGMKVPALGLWMRHNKCWVLEIFGLVVEIA
