+++ Sophia_Chernousova/resources/rosalind_revc.txt
@@ -1,0 +1,1 @@
+CGCAGTATTTGTGAGTTGGCGTTTCGGACCTGTGCAGCTAAAATATCCTGCCGGATCTTCACCTCACCAGATTATAGTCATAGAGCCCGAAGTTGTAAATTAATCGTCGTTCTGCTGCTCGTGACGACGACCTTATGTTGGGAGTGCTTTCTCCGTTAGAGTTTCCGCACTCTAAAAGAGGTACCACAAACGCGCCTCATTGGCTAAGTCAACCACCTTCTTCGAAGGGCTCCAGCTCGCGATCACCCGCTAGCAACTGGGTCCACTCGGCTTAATTACGTTCGCCTACATCCGTATTGAATTCCTGCGCGTGGTGCGCACGATCAGGCCGCTTTGAGTCGCCTTTGAAAAATAATTGGAGAGAAACAATCGACTATCGCAAGGAGATTGGCTCAGTCGGTCGACCCCGGACCACTACGTCCTACACATCGACTATAACCACACTTTTGCGCATCCCTTGTCTTGGAGTGGCTAATCCGAGGTCCGCAATCACGTCGTCCTATAATTCACCTCACTTGCTAGCTCAACCGCAAGTCGGTCTTTGTGTGCCTCTCCCAGAAGTTACTGCAATAAAACGGCAACTTGGAGGCGGCCTGGCCACTCGCCATTGCTTAAACAACGCTGTTGTGTTCGCTTACTTCCAAGGAACCCGCATGCTCGGTCCGACGTCAGCCTAATATGTTATAAGATAAGTGCACAATATCCGTATAACTGGGTAGTGGTGCCTTAAAATCACTTCAGAGTCGGCACACATCTATGTCGCGGCGATCCCTCGCGATGGAGTAGCCTTATGGACCATGGTGCGGGCCGCTAGTATTAGTATGAAGAATGAGGAAAGGGCCGTCGGACCGGCAGGATTAGCTTATGAAGTTCTTATGACCGCGCCTATCAATCTCCACGCTGTTCTTTCGAGGTAAAATAACCATGGAGCACGTTTTCTTGCAAGGCGTTCGTGTCTACTACATATCACCGACTAGACAAATATT
