+++ Sophia_Chernousova/resources/rosalind_rna.txt
@@ -1,0 +1,1 @@
+CTGACTCTGCCAATATGGGGCTGCGTCAAACATTCGTACAGTGAGTATCGCGGGCGATCCCATATTATGATGATAACATCTAGGTGGTGGTCGTGCAGAGATCTTTCACACCCTCCAGAGATGGACGGAGTATGACTGTGTGTACCATGCTTCTCTTCTTCGGACGCATGCCTCGTTTCGGCATCACCCTACATTACCCGTTCTCCTTAGGGGCGATCCCAGATAATGCGAACTTGGTTAAGCACTCCTAATACTGCGGCCCCTACTTGCGTTTCCGCTGTTGTTAAGTCCTTTCTACCCCTATCCAAGTAACGCCCGCCACTGATTTTGGAAACATTGGAGGGCACTGTTGTCATGTATCACAAGGCCTTCGAGCGCGGACCTCTCTGACGTGGTTGGTTGTAGATAGTGGACCATGTTTAACTGAGCGGCCGGATATCAGCTGCTGTCCACTGGTAAGCCGACGTCTCATAGACTTTTGACAGCTACTAAAGGCTCCACGTCCCGTGGGATAGCCCGTTGGTACTTTATAGAGAGTCCTGGCGTGTTGGATTGTTACCGATCAAGGATGCAGTACCATAGAAGGACTGTGTGATAAGTCTACCGACGTTGAACCGCAGCCTGATGTGATCTCTGATTCCCCCGCATCCCACGCTAGTGCTCCGAAAAATACCGGGACTTACCTATCGAATTCTACCCTATCCGGGGGATCTACGCGCGGTGTTCAACTAGGGCGAGCGAATCCCACAGGTCCTCACCACAATCCGCCTGTGGACGTGCACGTTCGTAGCTGGGGCCTACAGCGCATGGTTGGACAACAAGTGGCAATCTCTACCTCTAAGGCCCGCGTGACCCTTCAGGTCCCCGTAGCTTATCGGTGTCATAGCAGCGGTGCTACCCTCCGTTTGTTCCCCACTGCAACATCATGTATTA
