+++ Sophia_Chernousova/resources/rosalind_subs.txt
@@ -1,0 +1,2 @@
+ACATCGCTAAGAATGAATTGACTAAGAACTAAGAAATAGTCTAAGAAAGTCATGTACTAAGAAGATAGTTCTAAGAACTGCTAAGAACTAAGAACTAAGAACTAAGAATGAAGAGCGCTAAGAACTCCGGGCCTAAGAAGTCTAAGAACGGAACTAAGAACGTCTGACTCTAAGAAGCTCCCTAAGAACTAAGAACTAAGAAATCCTAAGAAGTCTAAGAACGCTACTGTGGACTAAGAACTAAGAAACTAAGAAAGACTAAGAACTAAGAAGACTAAGAAAACTAAGAACGCACTAAGAACTAAGAAACTAAGAAAGAGAGCTAAGAACCTAAGAACTAAGAAGCTAAGAATTCTAAGAACCTAAGAACCTAAGAAGCTAAGAATAACTAAGAAATTAACTAAGAAGCTCTAAGAATAGTCTAAGAACATGCTAAGAAAACATCCTAAGAACACTAAGAACTAAGAACTAAGAATTTCGTCTAAGAATCGCTCACTAAGAACTAAGAAACTAAGAACTAAGAACCCTAAGAAGAACTAAGAACTAAGAAGCCCCTAAGAATTTGCTTAGGGTGTCCTAAGAAATCTAAGAATCTAAGAATCTAAGAATCTAAGAATATCTAAGAATCACTAAGAAGCATAGTTACACTAAGAAACTCCTAAGAACTAAGAAGCTAAGAACTAAGAAATACTAAGAAGCCTAAGAAGGCCTAAGAAGAAAACTAAGAAAGCTAAGAAAGTCTAAGAAGAGCCCTAAGAAGGACCTAAGAATCCTAAGAACTAAGAATAGCTAAGAACTAAGAAGCTAAGAATGCAGCTGCTAAGAACTAAGAAACGTGTGCTAAGAACAAGCACTAAGAAGGTCTAAGAAGGTGACTAAGAAATCTCC
+CTAAGAACT
