--- resources/rosalind_dna.txt
@@ -1,1 +1,0 @@
-CGCTTTTCCATGACTGGGTCGAAAATGAGATTGATTCACCGCTTCATTATGTCGAGCTCCCATGAGTAGGAGTAACCCCGGTGAACTGAGTACAGCATTTGAAAGCTAGTCCGGTCGAGCGTGGTTGAACTCAAGAGATAGCGGGACAGGTAATTCCCGCTTGTTATCACCCAGGCCTGATCCTGTTGACGGGCTGCGGCCGGTTTACGTCTCGACCGTCTTCAGGGTCAATGTGCTATAAGTATAGGTGCATCAAAAGCACTGCATTTAAAGCGCTCGAAACACGAACAAAGGTGCACTGATGTATTTATCCACCAGGGCGAGCAACTCTTTAGAAATCGGCCCCATCAAATGAGTGCTAGCTGGACTTCACACTCGTTGGATGGACCACCTCCGTGGAGTTCGGACCGGGGAGCCGGTATGCTATCGTTATTTGTGGTACGCTCTACAGTGAGGATCCAATACAGGTTCCGGACATTCCCAAATCACGGAATCAGTGCAAGACATAACATCTTGCTGCCTTGAGGTAGGGACTGAACGCAAACTTGACGGGCCTCGGGGAACGGTGCTACAGTCCCATGATCCTGCAACGCTTACCGGGAGATTGAACTAGTGAGTAGGAACTCTAAGCGTCGTATAAGGTAGTCTTTAGGGACAGATTGGCATCGTATCAACGGTCGTGACCAATCGACTTACAAGAACCGTTCACTTGTCTCTTGTCAGGCGTTTGGTAAGTCTTCTCAGCCTGTGCGCGCACGCCGACTTCTTTCATAAACTTAAACGGGAGTAAATCGGAGATCGCTCTTGACTAAAGAACCTGGGGGT
