--- resources/rosalind_fib.txt
@@ -1,1 +1,0 @@
-35 4
