--- resources/rosalind_hamm.txt
@@ -1,2 +1,0 @@
-GGTAGGGGTGAGACTTTCGCAGGGTGTGAGGAGGCCAAGAGACACAAAGCCTCGGAGGATGCTGTCCGACTGTCGTAAGAGTAGAAGCGGCCAACAGTCACACAAACCACTGCGGCGACAAGAGGTTTCAATCCTCTTAGCGAGTCTCATAACCACATCAGGCCCTAAAGGTGCGCTTAACGTACAAACGCTAATACCGAATAGTAACGGAACAATGCAGATTGCAGTGAGCCGTTTTAGTTCCCGTGTCCCGGCGCTTGGTATTCTCTAACCATTGCTGAATATGACACCTCACTGAGTGCAGCACGAAATTCACATCGCCGCGGACGGGAATCGATCAACCGCCGCGAATTGAGTTTGTTCAGGACTTGCCGACAACAATAGAGCTATAGACAGGGGTATCAGACCACTCGGGGAATCCATCTCCCCAAACTCTCCTCCGAAGTTTGACGCTAGGTTCGGGTACGGGTTAGTAAACTCGTCTGCCAGTAAAGTGGTCCCTGCGCTCCCGCGTAAATGAGCCAGGTGTCTAATTCGTTTGGACCGCCAGCTCGTTGTCTTCTGGTAATGGCAAGCTTTTGTGCGCTCACGTGCTCATGGCGTCACGAATCTTCCCCCATTGGCATAAAGTGATGGAGAGAGAGGAAGTCTACAAACGGTCCGATGTTTTTTCGACTACAACCCATCGGCCGCCTGCCTAGAGACCACCCCACTCTGGGGCCCAGGGTCTTTACTGAATGCGGGACGCACGCGCAACACATCTGAGCATAGAGTACCTCGCTTTTAGCCATCAGGAGTTCCTGTTGGGGAACCCGTTAATTCCCAGACAAATTACATTCTTTTACTGTCAACCATTTTGCACTGGGACGCACCGGGTTTAAAGAGAACTGCCGCACTGCGGACTCGACGCCGCAATGCCCTACTTCCATCT
-GGTATATATGTCGCCTAAGACGGGATAGGGGGGAGCGAAGGTACCTAAGGCTGCCATGATACTGTCAGATGGGTCTAAACGCCGCAGCGGGCCTATTGCGTCCCGAGGTAACCGTCTTCAAGACGTTCAAATAGATCTAGCCTGACTCATCCGTCCACCAGGGCGCATACGTGCCCTAGGATTAGGGGGGCGGAACACGTAAACTCTGGGACAGTACATTCTTACCAATAGGCAATCAAGACCTAGCGGCCCGGTGGCGGATATTAAGTTGGCGTCGCTCTCTGCTGCGCTTCATTAACTGCGGCTCGGAATTCCACTCTTCGCGGGTGGCCGCCAGGAACTGGTGCAGAATTGAGTCATCTACCGAAGTACCGGATCAAAGAGAGCTGTCCACAGGGCCCTCAAAAAGCTACGGCGACCCTCATGTCATAACTTAGGGCCGTAGGTTGACGCTAGCACCGGGTGATGACCGGTCAGCTCAGCTGCGTCGTAAAAAAGCCAAGTGGTCACCAGTATCTTATACTGGAGTCCACTTGGTGTGGAACGGGAGTACAGTAGCTTCTAGAAATCGGGAGCGGGTCTGCGGTCAAGATGTGCTGGCGTTACTCCGCTTCCGCCGTTCACATATAAGTAGCGATCGAGAACCATTTTTACGTGAATCGGATGTTAAAGCGCCTGCAAGTCATTGCCCCCAACGATAGCAAGTGCGCAGCTCTAGGGCCTAGTGACTTTACCAAACGAGAGACTTAAGCGGAACAGACTTGAGCACGCGGTGCCTGGGATTCCGCTTTGACCGGTCCGTTTTGTGGGACCCGTCGAATCACACTAAAATTATAACGTTTTTCGGACCCCGATTGAAGACTGACACGCTACTGGTGCAAAGAGAATTGAAATCATGTGGTGTAGGCCCATCAATGCCTCTTTGCACTAT
