--- resources/rosalind_iprb.txt
@@ -1,1 +1,0 @@
-16 16 17
