--- resources/rosalind_perm.txt
@@ -1,1 +1,0 @@
-6
