--- resources/rosalind_prtm.txt
@@ -1,1 +1,0 @@
-NPWSCRFSQHWQEIAFLTLRLQELYGEDECMWRAIQEWDAGRCYYANPPHHMDARDCWLHSTQASVMKRMHAHKIEMGCCALHRCQPPMTYFVRVCQGTMWAWLELEGHLDCFPLRPPIGANYPMKYWLFFVQDQMIYCNEDTNGVIWPHWPYISKTVMKDSICHELNWGWSYWWDWPPTHFTVWAYTQVMLNWTKEELRDADEMDTQPLVWSHCRVCETEICKAHLAMLQYTDNLSHQFGGMMHYAAEKHAQTNESYDIDMHDIMLCAIQVYGLWDHLMFTYSIRPTPPKQATYFMTPCKPIEDMIISCCQSFVVSLRINMAYGKLNLGHCAPTTNSGFPNYGMNTRNAHWLEWLFDTEACQEHCRWAESICPKYCVASFRQPFNLGRAVYGYEGCDAVPAEGCHAFYAEMISWTGDCDQVIRRHWQCLEMNRWHKHAGRKDLAYCTYTWPKAPFGERNQCNERLDKGDAYVEFSRCANLARASLVSVAGNIKWKILQVCFIKRCMSSIYEQTVYRASQPDAMTSAPCWFWILTDQWGLTNETMYSWPYLPSWQKTEDQRASQEWGAGNAPTHKSLIQGMEGHSFKVSFLEENYRGHHSCAEVTPEDVNHTCGITKKEMAPPFSWWECKIPSFKLCELPWAEEHMELLFMMCLIWAWCCELKQDHWNHWESPMCRSFLPDAYCVWDVWHDPECQKIKQADIQGNHPATDIDAAEVYVRTTERQAWHSMEPTNKCIKYCKKIYVQWQASIEKSPHQPQCAHTYVTARMGIDAIKHQFVIQSEITMLHLKYTQECFKKTMLHVMIENISQGLFVAGWMFQWIHLNNYHDPLPSMRAKVHLAVWDASPFRCYAQNQKWYFPNNNVAFIFMWNEHWCTAVEQRDDNEPFKIPLCEMAQDRWCNRYFDYDGEFLVLFPNSTRGCQGMKVPALGLWMRHNKCWVLEIFGLVVEIA
