--- resources/rosalind_revc.txt
@@ -1,1 +1,0 @@
-CGCAGTATTTGTGAGTTGGCGTTTCGGACCTGTGCAGCTAAAATATCCTGCCGGATCTTCACCTCACCAGATTATAGTCATAGAGCCCGAAGTTGTAAATTAATCGTCGTTCTGCTGCTCGTGACGACGACCTTATGTTGGGAGTGCTTTCTCCGTTAGAGTTTCCGCACTCTAAAAGAGGTACCACAAACGCGCCTCATTGGCTAAGTCAACCACCTTCTTCGAAGGGCTCCAGCTCGCGATCACCCGCTAGCAACTGGGTCCACTCGGCTTAATTACGTTCGCCTACATCCGTATTGAATTCCTGCGCGTGGTGCGCACGATCAGGCCGCTTTGAGTCGCCTTTGAAAAATAATTGGAGAGAAACAATCGACTATCGCAAGGAGATTGGCTCAGTCGGTCGACCCCGGACCACTACGTCCTACACATCGACTATAACCACACTTTTGCGCATCCCTTGTCTTGGAGTGGCTAATCCGAGGTCCGCAATCACGTCGTCCTATAATTCACCTCACTTGCTAGCTCAACCGCAAGTCGGTCTTTGTGTGCCTCTCCCAGAAGTTACTGCAATAAAACGGCAACTTGGAGGCGGCCTGGCCACTCGCCATTGCTTAAACAACGCTGTTGTGTTCGCTTACTTCCAAGGAACCCGCATGCTCGGTCCGACGTCAGCCTAATATGTTATAAGATAAGTGCACAATATCCGTATAACTGGGTAGTGGTGCCTTAAAATCACTTCAGAGTCGGCACACATCTATGTCGCGGCGATCCCTCGCGATGGAGTAGCCTTATGGACCATGGTGCGGGCCGCTAGTATTAGTATGAAGAATGAGGAAAGGGCCGTCGGACCGGCAGGATTAGCTTATGAAGTTCTTATGACCGCGCCTATCAATCTCCACGCTGTTCTTTCGAGGTAAAATAACCATGGAGCACGTTTTCTTGCAAGGCGTTCGTGTCTACTACATATCACCGACTAGACAAATATT
