--- resources/rosalind_rna.txt
@@ -1,1 +1,0 @@
-CTGACTCTGCCAATATGGGGCTGCGTCAAACATTCGTACAGTGAGTATCGCGGGCGATCCCATATTATGATGATAACATCTAGGTGGTGGTCGTGCAGAGATCTTTCACACCCTCCAGAGATGGACGGAGTATGACTGTGTGTACCATGCTTCTCTTCTTCGGACGCATGCCTCGTTTCGGCATCACCCTACATTACCCGTTCTCCTTAGGGGCGATCCCAGATAATGCGAACTTGGTTAAGCACTCCTAATACTGCGGCCCCTACTTGCGTTTCCGCTGTTGTTAAGTCCTTTCTACCCCTATCCAAGTAACGCCCGCCACTGATTTTGGAAACATTGGAGGGCACTGTTGTCATGTATCACAAGGCCTTCGAGCGCGGACCTCTCTGACGTGGTTGGTTGTAGATAGTGGACCATGTTTAACTGAGCGGCCGGATATCAGCTGCTGTCCACTGGTAAGCCGACGTCTCATAGACTTTTGACAGCTACTAAAGGCTCCACGTCCCGTGGGATAGCCCGTTGGTACTTTATAGAGAGTCCTGGCGTGTTGGATTGTTACCGATCAAGGATGCAGTACCATAGAAGGACTGTGTGATAAGTCTACCGACGTTGAACCGCAGCCTGATGTGATCTCTGATTCCCCCGCATCCCACGCTAGTGCTCCGAAAAATACCGGGACTTACCTATCGAATTCTACCCTATCCGGGGGATCTACGCGCGGTGTTCAACTAGGGCGAGCGAATCCCACAGGTCCTCACCACAATCCGCCTGTGGACGTGCACGTTCGTAGCTGGGGCCTACAGCGCATGGTTGGACAACAAGTGGCAATCTCTACCTCTAAGGCCCGCGTGACCCTTCAGGTCCCCGTAGCTTATCGGTGTCATAGCAGCGGTGCTACCCTCCGTTTGTTCCCCACTGCAACATCATGTATTA
