--- resources/rosalind_subs.txt
@@ -1,2 +1,0 @@
-ACATCGCTAAGAATGAATTGACTAAGAACTAAGAAATAGTCTAAGAAAGTCATGTACTAAGAAGATAGTTCTAAGAACTGCTAAGAACTAAGAACTAAGAACTAAGAATGAAGAGCGCTAAGAACTCCGGGCCTAAGAAGTCTAAGAACGGAACTAAGAACGTCTGACTCTAAGAAGCTCCCTAAGAACTAAGAACTAAGAAATCCTAAGAAGTCTAAGAACGCTACTGTGGACTAAGAACTAAGAAACTAAGAAAGACTAAGAACTAAGAAGACTAAGAAAACTAAGAACGCACTAAGAACTAAGAAACTAAGAAAGAGAGCTAAGAACCTAAGAACTAAGAAGCTAAGAATTCTAAGAACCTAAGAACCTAAGAAGCTAAGAATAACTAAGAAATTAACTAAGAAGCTCTAAGAATAGTCTAAGAACATGCTAAGAAAACATCCTAAGAACACTAAGAACTAAGAACTAAGAATTTCGTCTAAGAATCGCTCACTAAGAACTAAGAAACTAAGAACTAAGAACCCTAAGAAGAACTAAGAACTAAGAAGCCCCTAAGAATTTGCTTAGGGTGTCCTAAGAAATCTAAGAATCTAAGAATCTAAGAATCTAAGAATATCTAAGAATCACTAAGAAGCATAGTTACACTAAGAAACTCCTAAGAACTAAGAAGCTAAGAACTAAGAAATACTAAGAAGCCTAAGAAGGCCTAAGAAGAAAACTAAGAAAGCTAAGAAAGTCTAAGAAGAGCCCTAAGAAGGACCTAAGAATCCTAAGAACTAAGAATAGCTAAGAACTAAGAAGCTAAGAATGCAGCTGCTAAGAACTAAGAAACGTGTGCTAAGAACAAGCACTAAGAAGGTCTAAGAAGGTGACTAAGAAATCTCC
-CTAAGAACT
